Feat: Connent End Button and Quit Code
